Dar 331 book management (#55)
* Add book icon to Navbar and remove video icon

Replaced the video icon with the book icon in the Navbar component and added a new entry for books in the navigation items. Also introduced a new Svelte file for the book icon’s SVG representation.

* feat: add book management

* Update book pagination and add bulk delete functionality

Increased the default pagination limit to 24 on the server side. Added translations for book-related terms, and introduced bulk delete and individual delete functionalities with user feedback in the admin book page.